requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioresponses==0.7.8
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
discord.py>=2.4
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
gunicorn==21.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai>=0.5.0
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.2
packaging==24.2
Pillow>=10
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil>=2.9.0
python-dotenv>=1.0
requests==2.32.3
rsa==4.9
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
openai>=1.58.0

Werkzeug==3.1.3
yarl==1.18.3
rapidfuzz>=3.0
dateparser==1.2.0
yt-dlp>=2024.12.23
